services/token/requirements.txt
bcrypt==3.1.7
Flask==1.1.1
Flask-API==2.0
flask-expects-json==1.4.0
Flask-PyMongo==2.3.0
PyJWT==1.7.1
pymongo==3.10.0
requests==2.22.0
